@@ -91,7 +91,7 @@ public:
    * @details Times are in µs, fidelities are in [0,1].
    */
   struct OperationProperties {
-    Scope                    scope; // local or global
+    Scope scope;                    // local or global
     std::unordered_set<Zone> zones; // the zones where the gate can be applied
     Value time;     // the time the gate takes to be applied in µs
     Value fidelity; // the fidelity of the gate
@@ -103,26 +103,26 @@ public:
    * µm/µs.
    */
   struct ShuttlingProperties {
-    Index  rows          = 0; // maximum number of rows in one AOD
-    Index  cols          = 0; // maximum number of columns in one AOD
-    Number minX          = 0; // minimum x position of the AOD
-    Number maxX          = 0; // maximum x position of the AOD
-    Number minY          = 0; // minimum y position of the AOD
-    Number maxY          = 0; // maximum y position of the AOD
-    Value  speed         = 0; // speed of the AOD in µm/µs
-    Value  fidelity      = 1; // fidelity during the shuttling
-    Value  loadTime      = 0; // time to activate the AOD in µs
-    Value  loadFidelity  = 1; // fidelity of the load
-    Value  storeTime     = 0; // time to deactivate the AOD in µs
-    Value  storeFidelity = 1; // fidelity of the store
+    Index rows = 0;          // maximum number of rows in one AOD
+    Index cols = 0;          // maximum number of columns in one AOD
+    Number minX = 0;         // minimum x position of the AOD
+    Number maxX = 0;         // maximum x position of the AOD
+    Number minY = 0;         // minimum y position of the AOD
+    Number maxY = 0;         // maximum y position of the AOD
+    Value speed = 0;         // speed of the AOD in µm/µs
+    Value fidelity = 1;      // fidelity during the shuttling
+    Value loadTime = 0;      // time to activate the AOD in µs
+    Value loadFidelity = 1;  // fidelity of the load
+    Value storeTime = 0;     // time to deactivate the AOD in µs
+    Value storeFidelity = 1; // fidelity of the store
   };
   struct ZoneProperties {
-    std::string name;         // the name of the zone
-    Number      minX     = 0; // minimum x dimension
-    Number      maxX     = 0; // maximum x dimension
-    Number      minY     = 0; // minimum y dimension
-    Number      maxY     = 0; // maximum y dimension
-    Value       fidelity = 1; // fidelity during idling
+    std::string name;   // the name of the zone
+    Number minX = 0;    // minimum x dimension
+    Number maxX = 0;    // maximum x dimension
+    Number minY = 0;    // minimum y dimension
+    Number maxY = 0;    // maximum y dimension
+    Value fidelity = 1; // fidelity during idling
   };
 
 protected:
@@ -134,7 +134,7 @@ protected:
       gateSet; // all possible operations by their type, i.e. gate set
   DecoherenceTimes decoherenceTimes;          // the decoherence characteristic
   std::vector<ShuttlingProperties> shuttling; // all properties regarding AODs
-  Distance                         minAtomDistance =
+  Distance minAtomDistance =
       0; // minimal distance that must be kept between atoms
   Distance interactionRadius = 0; // the Rydberg radius
   Distance noInteractionRadius =
@@ -190,8 +190,8 @@ public:
   [[nodiscard]] auto getNoInteractionRadius() const -> Distance {
     return noInteractionRadius;
   }
-  [[nodiscard]] auto
-  getPropertiesOfZone(const Zone& zone) const -> const ZoneProperties& {
+  [[nodiscard]] auto getPropertiesOfZone(const Zone& zone) const
+      -> const ZoneProperties& {
     return zones[zone];
   }
   [[nodiscard]] auto getPropertiesOfOperation(const FullOpType& t) const
@@ -210,8 +210,8 @@ public:
    * @param j address of second site
    * @return the distance in µm
    */
-  [[nodiscard]] auto getDistance(const Index& i,
-                                 const Index& j) const -> Index {
+  [[nodiscard]] auto getDistance(const Index& i, const Index& j) const
+      -> Index {
     return (getPositionOfSite(j) - getPositionOfSite(i)).length();
   }
   [[nodiscard]] auto getZoneAt(const Point& p) const -> Zone;
@@ -222,17 +222,17 @@ public:
   [[nodiscard]] auto isAllowedLocally(const FullOpType& t) const -> bool;
   /// Checks whether the gate can be applied (locally) in this zone.
   [[nodiscard]] auto isAllowedLocally(const FullOpType& t,
-                                      const Zone&       zone) const -> bool;
+                                      const Zone& zone) const -> bool;
   /// Checks whether the gate can be applied (locally) on this qubit.
   [[nodiscard]] auto isAllowedLocallyAt(const FullOpType& t,
-                                        const Point&      p) const -> bool;
+                                        const Point& p) const -> bool;
   /// Checks whether the gate is a global gate for this Zone.
   [[nodiscard]] auto isAllowedGlobally(const FullOpType& t) const -> bool;
   [[nodiscard]] auto isAllowedGlobally(const FullOpType& t,
-                                       const Zone&       zone) const -> bool;
+                                       const Zone& zone) const -> bool;
   [[nodiscard]] auto getNrowsInZone(const Zone& z) const -> Index;
-  [[nodiscard]] auto
-  getSitesInRow(const Zone& z, const Index& row) const -> std::vector<Index>;
+  [[nodiscard]] auto getSitesInRow(const Zone& z, const Index& row) const
+      -> std::vector<Index>;
   [[nodiscard]] auto getNearestXLeft(const Number& x, const Zone& z,
                                      bool proper = true) const -> Number;
   [[nodiscard]] auto getNearestXRight(const Number& x, const Zone& z,
@@ -249,34 +249,35 @@ public:
   [[nodiscard]] auto hasSiteDown(const Point& p, bool proper = false,
                                  bool sameZone = false) const
       -> std::pair<std::vector<Point>::const_iterator, bool>;
-  [[nodiscard]] auto
-  getNearestSiteLeft(const Point& p, bool proper = false,
-                     bool sameZone = false) const -> std::optional<Index>;
-  [[nodiscard]] auto
-  getNearestSiteRight(const Point& p, bool proper = false,
-                      bool sameZone = false) const -> std::optional<Index>;
-  [[nodiscard]] auto
-  getNearestSiteUp(const Point& p, bool proper = false,
-                   bool sameZone = false) const -> std::optional<Index>;
-  [[nodiscard]] auto
-  getNearestSiteDown(const Point& p, bool proper = false,
-                     bool sameZone = false) const -> std::optional<Index>;
-  [[nodiscard]] auto
-  getNearestSiteUpRight(const Point& p, bool proper = false,
-                        bool sameZone = false) const -> std::optional<Index>;
-  [[nodiscard]] auto
-  getNearestSiteUpLeft(const Point& p, bool proper = false,
-                       bool sameZone = false) const -> std::optional<Index>;
-  [[nodiscard]] auto
-  getNearestSiteDownLeft(const Point& p, bool proper = false,
-                         bool sameZone = false) const -> std::optional<Index>;
-  [[nodiscard]] auto
-  getNearestSiteDownRight(const Point& p, bool proper = false,
-                          bool sameZone = false) const -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteLeft(const Point& p, bool proper = false,
+                                        bool sameZone = false) const
+      -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteRight(const Point& p, bool proper = false,
+                                         bool sameZone = false) const
+      -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteUp(const Point& p, bool proper = false,
+                                      bool sameZone = false) const
+      -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteDown(const Point& p, bool proper = false,
+                                        bool sameZone = false) const
+      -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteUpRight(const Point& p, bool proper = false,
+                                           bool sameZone = false) const
+      -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteUpLeft(const Point& p, bool proper = false,
+                                          bool sameZone = false) const
+      -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteDownLeft(const Point& p, bool proper = false,
+                                            bool sameZone = false) const
+      -> std::optional<Index>;
+  [[nodiscard]] auto getNearestSiteDownRight(const Point& p,
+                                             bool proper = false,
+                                             bool sameZone = false) const
+      -> std::optional<Index>;
   [[nodiscard]] auto getSiteAt(const Point& p) const -> std::optional<Index>;
   [[nodiscard]] auto getSitesInZone(const Zone& z) const -> std::vector<Index>;
-  [[nodiscard]] auto
-  withConfig(const Configuration& config) const -> Architecture;
+  [[nodiscard]] auto withConfig(const Configuration& config) const
+      -> Architecture;
   [[nodiscard]] auto getPositionOffsetBy(const Point& p, const Number& rows,
                                          const Number& cols) const -> Point;
 
