@@ -11,6 +11,7 @@
 #include <cstdint>
 #include <istream>
 #include <sstream>
+#include <stdexcept>
 #include <string>
 #include <unordered_map>
 namespace na {
@@ -22,8 +23,8 @@ enum class NAMappingMethod : std::uint8_t {
 static const std::unordered_map<std::string, NAMappingMethod> STRING_TO_METHOD =
     {{"naive", NAMappingMethod::Naive},
      {"maximize parallelism", NAMappingMethod::MaximizeParallelismHeuristic}};
-[[nodiscard]] inline auto
-getMethodOfString(const std::string& method) -> NAMappingMethod {
+[[nodiscard]] inline auto getMethodOfString(const std::string& method)
+    -> NAMappingMethod {
   std::string methodLowerCase = method;
   std::transform(methodLowerCase.begin(), methodLowerCase.end(),
                  methodLowerCase.begin(),
@@ -38,9 +39,9 @@ getMethodOfString(const std::string& method) -> NAMappingMethod {
 }
 class Configuration {
 private:
-  std::size_t     patchRows = 1;
-  std::size_t     patchCols = 1;
-  NAMappingMethod method    = NAMappingMethod::MaximizeParallelismHeuristic;
+  std::size_t patchRows = 1;
+  std::size_t patchCols = 1;
+  NAMappingMethod method = NAMappingMethod::MaximizeParallelismHeuristic;
 
 public:
   Configuration() = default;
