@@ -20,8 +20,8 @@
 
 namespace na {
 
-using Color            = std::uint16_t;
-using Edge             = std::pair<qc::Qubit, qc::Qubit>;
+using Color = std::uint16_t;
+using Edge = std::pair<qc::Qubit, qc::Qubit>;
 using InteractionGraph = qc::Layer::InteractionGraph;
 
 class NAGraphAlgorithms {
@@ -48,7 +48,7 @@ public:
    * @return a set of edges with at least one node in vs
    */
   [[nodiscard]] static auto
-  coveredEdges(const InteractionGraph&              g,
+  coveredEdges(const InteractionGraph& g,
                const std::unordered_set<qc::Qubit>& vs)
       -> std::unordered_set<Edge, qc::PairHash<qc::Qubit, qc::Qubit>>;
 
@@ -82,15 +82,15 @@ public:
    * @param partialOrder the partial order of the fixed vertices induced by the
    * coloring so far
    * @param ranks the ranks of the vertices in the sequence, i.e., a mapping of
-   * nodes in the sequence to their index in the seuqence, i.e., their rank
+   * nodes in the sequence to their index in the sequence, i.e., their rank
    * @return the least admissible color
    */
   [[nodiscard]] static auto getLeastAdmissibleColor(
       const std::unordered_map<Edge, Color, qc::PairHash<qc::Qubit, qc::Qubit>>&
-                   coloring,
+          coloring,
       const Color& maxColor, const Edge& e, const qc::Qubit& v,
-      const std::vector<qc::Qubit>&                             sequence,
-      const qc::DirectedAcyclicGraph<qc::Qubit>&                partialOrder,
+      const std::vector<qc::Qubit>& sequence,
+      const qc::DirectedAcyclicGraph<qc::Qubit>& partialOrder,
       const std::unordered_map<std::pair<qc::Qubit, Color>, std::size_t,
                                qc::PairHash<qc::Qubit, Color>>& ranks) -> Color;
 
@@ -103,7 +103,7 @@ public:
    * @return a valid coloring of the edges
    */
   [[nodiscard]] static auto colorEdges(
-      const InteractionGraph&                                             g,
+      const InteractionGraph& g,
       const std::unordered_set<Edge, qc::PairHash<qc::Qubit, qc::Qubit>>& edges,
       const std::vector<qc::Qubit>& nodesQueue)
       -> std::pair<
@@ -130,11 +130,12 @@ public:
    * @brief Groups the vertices in the sequence by connected components.
    * @param g the graph with connected components
    * @param sequence the sequence of vertices
-   * @return a new seuqence with the vertices grouped by connected components
+   * @return a new sequence with the vertices grouped by connected components
    */
-  [[nodiscard]] static auto groupByConnectedComponent(
-      const InteractionGraph&       g,
-      const std::vector<qc::Qubit>& sequence) -> std::vector<qc::Qubit>;
+  [[nodiscard]] static auto
+  groupByConnectedComponent(const InteractionGraph& g,
+                            const std::vector<qc::Qubit>& sequence)
+      -> std::vector<qc::Qubit>;
 
   /**
    * @brief The result will be (1) a mapping of vertices to their relative x
@@ -148,11 +149,13 @@ public:
    * 3. Find relative x-positions for the fixed vertices.
    * 4. Compute the resting positions of the moveable vertices between the fixed
    * vertices.
-   * @param g the interaction graph
+   * @param g        the interaction graph
+   * @param maxSites the number of sites horizontally in the entnagling zones
    * @return A tuple of vectors containing the moveable vertices, the fixed
    * vertices mapped to their relative x-position in every time step.
    */
-  [[nodiscard]] static auto computeSequence(const InteractionGraph& g)
+  [[nodiscard]] static auto computeSequence(const InteractionGraph& g,
+                                            std::size_t maxSites)
       -> std::pair<std::vector<std::unordered_map<qc::Qubit, std::int64_t>>,
                    std::unordered_map<qc::Qubit, std::int64_t>>;
 };
