@@ -2,7 +2,9 @@
 
 #include "Logic.hpp"
 
+#include <algorithm>
 #include <cmath>
+#include <cstddef>
 #include <cstdint>
 #include <initializer_list>
 #include <limits>
@@ -710,6 +712,8 @@ bool TermDepthComparator::operator()(const LogicTerm& t1,
       t1.getDepth() == t2.getDepth()) {
     return t1.getID() > t2.getID();
   }
-  { return t1.getDepth() > t2.getDepth(); }
+  {
+    return t1.getDepth() > t2.getDepth();
+  }
 }
 } // namespace logicbase
