@@ -4,11 +4,12 @@
 // information.
 //
 
-#include "Architecture.hpp"
+#include "na/Architecture.hpp"
 
-#include "Configuration.hpp"
+#include "Definitions.hpp"
+#include "ir/operations/OpType.hpp"
+#include "na/Configuration.hpp"
 #include "na/NADefinitions.hpp"
-#include "operations/OpType.hpp"
 
 #include <algorithm>
 #include <cstdint>
@@ -20,6 +21,7 @@
 #include <limits>
 #include <map>
 #include <nlohmann/json.hpp>
+#include <nlohmann/json_fwd.hpp>
 #include <optional>
 #include <sstream>
 #include <stdexcept>
@@ -30,8 +32,8 @@
 
 namespace na {
 
-auto Architecture::fromFile(const std::string& jsonFn,
-                            const std::string& csvFn) -> void {
+auto Architecture::fromFile(const std::string& jsonFn, const std::string& csvFn)
+    -> void {
   std::ifstream jsonS(jsonFn);
   if (!jsonS.good()) {
     std::stringstream ss;
@@ -47,9 +49,9 @@ auto Architecture::fromFile(const std::string& jsonFn,
   *this = Architecture(jsonS, csvS);
 }
 
-auto Architecture::fromFileStream(std::istream& jsonS,
-                                  std::istream& csvS) -> void {
-  nlohmann::json data;
+auto Architecture::fromFileStream(std::istream& jsonS, std::istream& csvS)
+    -> void {
+  nlohmann::basic_json data;
   // load CSV
   // auxiliary variables
   std::string line;
@@ -59,8 +61,8 @@ auto Architecture::fromFileStream(std::istream& jsonS,
     while (csvS >> line) {
       ++lineno;
       std::stringstream lineStream(line); // make a stream of this line
-      std::string       sX;
-      std::string       sY;
+      std::string sX;
+      std::string sY;
 
       std::getline(lineStream, sX, ',');
       std::getline(lineStream, sY);
@@ -98,17 +100,17 @@ auto Architecture::fromFileStream(std::istream& jsonS,
       initialZones.emplace_back(nameToZone.find(zone)->second);
     }
     for (auto const& op : data["operations"]) {
-      const std::string        opName = op["name"];
-      const FullOpType         ty     = {qc::opTypeFromString(opName),
-                                         opName.find_first_not_of('c')};
-      const Scope              sc     = getScopeOfString(op["type"]);
-      std::unordered_set<Zone> zo     = {};
+      const std::string opName = op["name"];
+      const FullOpType ty = {qc::opTypeFromString(opName),
+                             opName.find_first_not_of('c')};
+      const Scope sc = getScopeOfString(op["type"]);
+      std::unordered_set<Zone> zo = {};
       for (auto const& zs : op["zones"]) {
         zo.emplace(nameToZone.find(zs)->second);
       }
-      const Value               fi = op["fidelity"];
-      const Value               ti = op["time"];
-      OperationProperties const o  = {sc, zo, ti, fi};
+      const Value fi = op["fidelity"];
+      const Value ti = op["time"];
+      OperationProperties const o = {sc, zo, ti, fi};
       gateSet.emplace(ty, o);
     }
     decoherenceTimes = {data["decoherence"]["t1"], data["decoherence"]["t2"]};
@@ -122,8 +124,8 @@ auto Architecture::fromFileStream(std::istream& jsonS,
           sh["store"]["time"], sh["store"]["fidelity"]};
       shuttling.emplace_back(sp);
     }
-    minAtomDistance     = data["minAtomDistance"];
-    interactionRadius   = data["interactionRadius"];
+    minAtomDistance = data["minAtomDistance"];
+    interactionRadius = data["interactionRadius"];
     noInteractionRadius = data["noInteractionRadius"];
   } catch (std::exception& e) {
     throw std::runtime_error(
@@ -150,19 +152,22 @@ auto Architecture::isAllowedLocally(const FullOpType& t) const -> bool {
   const auto it = gateSet.find(t);
   return it != gateSet.end() && it->second.scope == Scope::Local;
 }
-auto Architecture::isAllowedLocally(const FullOpType& t,
-                                    const Zone&       zone) const -> bool {
+auto Architecture::isAllowedLocally(const FullOpType& t, const Zone& zone) const
+    -> bool {
   if (!isAllowedLocally(t)) {
     return false; // gate not supported at all
   }
-  const auto  it        = gateSet.find(t);
+  const auto it = gateSet.find(t);
+  if (it == gateSet.end()) {
+    qc::unreachable(); // please the clang-tidy null dereference checker
+  }
   const auto& gateZones = it->second.zones;
   // zone exists in gateZones
   return gateZones.find(zone) != gateZones.end();
 }
 
-auto Architecture::isAllowedLocallyAt(const FullOpType& t,
-                                      const Point&      p) const -> bool {
+auto Architecture::isAllowedLocallyAt(const FullOpType& t, const Point& p) const
+    -> bool {
   const auto& it =
       std::find_if(zones.cbegin(), zones.cend(), [&](const auto& zProp) {
         return p.x >= zProp.minX && p.x <= zProp.maxX && p.y >= zProp.minY &&
@@ -183,11 +188,14 @@ auto Architecture::isAllowedGlobally(const FullOpType& t) const -> bool {
 }
 
 auto Architecture::isAllowedGlobally(const FullOpType& t,
-                                     const Zone&       zone) const -> bool {
+                                     const Zone& zone) const -> bool {
   if (!isAllowedGlobally(t)) {
     return false; // gate not supported at all
   }
-  const auto  it        = gateSet.find(t);
+  const auto it = gateSet.find(t);
+  if (it == gateSet.end()) {
+    qc::unreachable(); // please the clang-tidy null dereference checker
+  }
   const auto& gateZones = it->second.zones;
   // zone exists in gateZones
   return gateZones.find(zone) != gateZones.end();
@@ -219,9 +227,9 @@ auto Architecture::getColsInZone(const Zone& z) const -> std::vector<Number> {
 auto Architecture::getNrowsInZone(const Zone& z) const -> Index {
   return Architecture::getRowsInZone(z).size();
 }
-auto Architecture::getSitesInRow(const Zone&  z,
-                                 const Index& row) const -> std::vector<Index> {
-  const auto         y = Architecture::getRowsInZone(z)[row];
+auto Architecture::getSitesInRow(const Zone& z, const Index& row) const
+    -> std::vector<Index> {
+  const auto y = Architecture::getRowsInZone(z)[row];
   std::vector<Index> atoms;
   for (Index i = 0; i < sites.size(); ++i) {
     const auto& s = sites[i];
@@ -285,8 +293,9 @@ auto Architecture::hasSiteLeft(const Point& p, bool proper, bool sameZone) const
       });
   return {it, it != sites.crend()};
 }
-auto Architecture::hasSiteRight(const Point& p, bool proper, bool sameZone)
-    const -> std::pair<std::vector<Point>::const_iterator, bool> {
+auto Architecture::hasSiteRight(const Point& p, bool proper,
+                                bool sameZone) const
+    -> std::pair<std::vector<Point>::const_iterator, bool> {
   const auto& zone = getZoneAt(p);
   const auto& it =
       std::find_if(sites.cbegin(), sites.cend(), [&](const auto& s) {
@@ -356,7 +365,7 @@ auto Architecture::getNearestSiteDown(const Point& p, const bool proper,
 auto Architecture::getNearestSiteUpRight(const Point& p, const bool proper,
                                          const bool sameZone) const
     -> std::optional<Index> {
-  const auto&          zone = getZoneAt(p);
+  const auto& zone = getZoneAt(p);
   std::optional<Index> opt;
   for (std::size_t i = 0; i < sites.size(); ++i) {
     const auto& s = sites[i];
@@ -372,7 +381,7 @@ auto Architecture::getNearestSiteUpRight(const Point& p, const bool proper,
 auto Architecture::getNearestSiteUpLeft(const Point& p, const bool proper,
                                         const bool sameZone) const
     -> std::optional<Index> {
-  const auto&          zone = getZoneAt(p);
+  const auto& zone = getZoneAt(p);
   std::optional<Index> opt;
   for (std::size_t i = 0; i < sites.size(); ++i) {
     const auto& s = sites[i];
@@ -388,7 +397,7 @@ auto Architecture::getNearestSiteUpLeft(const Point& p, const bool proper,
 auto Architecture::getNearestSiteDownLeft(const Point& p, const bool proper,
                                           const bool sameZone) const
     -> std::optional<Index> {
-  const auto&          zone = getZoneAt(p);
+  const auto& zone = getZoneAt(p);
   std::optional<Index> opt;
   for (std::size_t i = 0; i < sites.size(); ++i) {
     const auto& s = sites[i];
@@ -404,7 +413,7 @@ auto Architecture::getNearestSiteDownLeft(const Point& p, const bool proper,
 auto Architecture::getNearestSiteDownRight(const Point& p, const bool proper,
                                            const bool sameZone) const
     -> std::optional<Index> {
-  const auto&          zone = getZoneAt(p);
+  const auto& zone = getZoneAt(p);
   std::optional<Index> opt;
   for (std::size_t i = 0; i < sites.size(); ++i) {
     const auto& s = sites[i];
@@ -439,7 +448,7 @@ auto Architecture::withConfig(const Configuration& config) const
         if (!usedSites[i]) {
           // s is not used yet
           bool patchFittable = true;
-          auto row           = i;
+          auto row = i;
           for (std::size_t r = 1; r < config.getPatchRows(); ++r) {
             auto col = row;
             for (std::size_t c = 1; c < config.getPatchCols(); ++c) {
@@ -533,11 +542,11 @@ auto Architecture::getPositionOffsetBy(const Point& p, const Number& rows,
 
   // get position of nearest site
   const auto nearestSitePos = getPositionOfSite(nearestSiteOpt.value());
-  const auto dx             = p.x - nearestSitePos.x;
-  const auto dy             = p.y - nearestSitePos.y;
-  auto       anchorSitePos  = nearestSitePos;
-  auto       r              = std::abs(rows);
-  auto       c              = std::abs(cols);
+  const auto dx = p.x - nearestSitePos.x;
+  const auto dy = p.y - nearestSitePos.y;
+  auto anchorSitePos = nearestSitePos;
+  auto r = std::abs(rows);
+  auto c = std::abs(cols);
 
   for (; r > 0; --r) {
     auto nextSiteOpt = rows >= 0 ? getNearestSiteDown(anchorSitePos, true, true)
@@ -545,7 +554,14 @@ auto Architecture::getPositionOffsetBy(const Point& p, const Number& rows,
     if (!nextSiteOpt.has_value()) {
       break;
     }
-    anchorSitePos = getPositionOfSite(nextSiteOpt.value());
+    const auto& newAnchorSitePos = getPositionOfSite(nextSiteOpt.value());
+    // The following check is used to decide whether the patch is located
+    // outside of the zone Patches that overlap the border of the zone are not
+    // possible with this approach yet
+    if (std::abs(newAnchorSitePos.y - anchorSitePos.y) < std::abs(dy)) {
+      break;
+    }
+    anchorSitePos = newAnchorSitePos;
   }
 
   for (; c > 0; --c) {
@@ -555,7 +571,14 @@ auto Architecture::getPositionOffsetBy(const Point& p, const Number& rows,
     if (!nextSiteOpt.has_value()) {
       break;
     }
-    anchorSitePos = getPositionOfSite(nextSiteOpt.value());
+    const auto& newAnchorSitePos = getPositionOfSite(nextSiteOpt.value());
+    // The following check is used to decide whether the patch is located
+    // outside of the zone Patches that overlap the border of the zone are not
+    // possible with this approach yet
+    if (std::abs(newAnchorSitePos.x - anchorSitePos.x) < std::abs(dx)) {
+      break;
+    }
+    anchorSitePos = newAnchorSitePos;
   }
   anchorSitePos.x =
       cols >= 0 ? anchorSitePos.x + c * d : anchorSitePos.x - c * d;
