@@ -3,19 +3,19 @@
 // See README.md or go to https://github.com/cda-tum/qmap for more information.
 //
 
-#include "Architecture.hpp"
-#include "Configuration.hpp"
 #include "Definitions.hpp"
-#include "NAMapper.hpp"
-#include "QuantumComputation.hpp"
 #include "datastructures/Layer.hpp"
+#include "ir/QuantumComputation.hpp"
+#include "ir/operations/CompoundOperation.hpp"
+#include "ir/operations/OpType.hpp"
+#include "ir/operations/StandardOperation.hpp"
+#include "na/Architecture.hpp"
+#include "na/Configuration.hpp"
 #include "na/NAComputation.hpp"
+#include "na/NAMapper.hpp"
 #include "na/operations/NAGlobalOperation.hpp"
 #include "na/operations/NALocalOperation.hpp"
 #include "na/operations/NAShuttlingOperation.hpp"
-#include "operations/CompoundOperation.hpp"
-#include "operations/OpType.hpp"
-#include "operations/StandardOperation.hpp"
 
 #include <algorithm>
 #include <cstddef>
@@ -30,7 +30,9 @@
 
 namespace na {
 auto validateAODConstraints(const NAComputation& comp) -> bool {
+  std::size_t counter = 1; // the first operation is `init at ...;`
   for (const auto& naOp : comp) {
+    ++counter;
     if (naOp->isShuttlingOperation()) {
       const auto& shuttlingOp =
           dynamic_cast<const NAShuttlingOperation&>(*naOp);
@@ -44,27 +46,43 @@ auto validateAODConstraints(const NAComputation& comp) -> bool {
           const auto& e1 = shuttlingOp.getEnd()[i];
           const auto& e2 = shuttlingOp.getEnd()[j];
           if (*s1 == *s2) {
+            std::cout << "Error in op number " << counter
+                      << " (two start points identical)\n";
             return false;
           }
           if (*e1 == *e2) {
+            std::cout << "Error in op number " << counter
+                      << " (two end points identical)\n";
             return false;
           }
           if (s1->x == s2->x && e1->x != e2->x) {
+            std::cout << "Error in op number " << counter
+                      << " (columns not preserved)\n";
             return false;
           }
           if (s1->y == s2->y && e1->y != e2->y) {
+            std::cout << "Error in op number " << counter
+                      << " (rows not preserved)\n";
             return false;
           }
           if (s1->x < s2->x && e1->x >= e2->x) {
+            std::cout << "Error in op number " << counter
+                      << " (column order not preserved)\n";
             return false;
           }
           if (s1->y < s2->y && e1->y >= e2->y) {
+            std::cout << "Error in op number " << counter
+                      << " (row order not preserved)\n";
             return false;
           }
           if (s1->x > s2->x && e1->x <= e2->x) {
+            std::cout << "Error in op number " << counter
+                      << " (column order not preserved)\n";
             return false;
           }
           if (s1->y > s2->y && e1->y <= e2->y) {
+            std::cout << "Error in op number " << counter
+                      << " (row order not preserved)\n";
             return false;
           }
         }
@@ -76,6 +94,8 @@ auto validateAODConstraints(const NAComputation& comp) -> bool {
           const auto& a = localOp.getPositions()[i];
           const auto& b = localOp.getPositions()[j];
           if (*a == *b) {
+            std::cout << "Error in op number " << counter
+                      << " (identical positions)\n";
             return false;
           }
         }
@@ -86,10 +106,10 @@ auto validateAODConstraints(const NAComputation& comp) -> bool {
 }
 
 auto retrieveQuantumComputation(const NAComputation& nac,
-                                const Architecture&  arch)
+                                const Architecture& arch)
     -> qc::QuantumComputation {
-  qc::QuantumComputation               qComp(nac.getInitialPositions().size());
-  std::vector<Point>                   positionOfQubits;
+  qc::QuantumComputation qComp(nac.getInitialPositions().size());
+  std::vector<Point> positionOfQubits;
   std::unordered_map<Point, qc::Qubit> positionToQubit;
   positionOfQubits.reserve(nac.getInitialPositions().size());
   qc::Qubit n = 0;
@@ -160,15 +180,15 @@ auto retrieveQuantumComputation(const NAComputation& nac,
 }
 
 auto checkEquivalence(const qc::QuantumComputation& circ,
-                      const NAComputation&          nac,
-                      const Architecture&           arch) -> bool {
-  auto            naQComp = retrieveQuantumComputation(nac, arch);
+                      const NAComputation& nac, const Architecture& arch)
+    -> bool {
+  auto naQComp = retrieveQuantumComputation(nac, arch);
   const qc::Layer qLayer(circ);
-  int             line = 0;
+  int line = 0;
   for (const auto& op : naQComp) {
     ++line;
     const auto& executableSet = qLayer.getExecutableSet();
-    const auto& it            = std::find_if(
+    const auto& it = std::find_if(
         executableSet.begin(), executableSet.end(),
         [&op](const std::shared_ptr<qc::Layer::DAGVertex>& vertex) {
           return *vertex->getOperation() == *op;
@@ -294,7 +314,7 @@ TEST(NAMapper, Exceptions) {
           }
       ]
   })");
-  std::stringstream  gridSS;
+  std::stringstream gridSS;
   gridSS << "x,y\n";
   // entangling zone (4 x 36 = 144 sites)
   for (std::size_t y = 0; y <= 36; y += 12) {
@@ -448,7 +468,7 @@ TEST(NAMapper, QAOA10) {
           }
       ]
   })");
-  std::stringstream  gridSS;
+  std::stringstream gridSS;
   gridSS << "x,y\n";
   // entangling zone (4 x 36 = 144 sites)
   for (std::size_t y = 0; y <= 36; y += 12) {
@@ -729,8 +749,8 @@ rz(3.9927041) q[3];
 rz(3.9927041) q[4];
 rz(3.9927041) q[5];
 rz(3.9927041) q[7];)";
-  const auto&       circ = qc::QuantumComputation::fromQASM(qasm);
-  const auto&       arch = na::Architecture(archIS, gridSS);
+  const auto& circ = qc::QuantumComputation::fromQASM(qasm);
+  const auto& arch = na::Architecture(archIS, gridSS);
   // ---------------------------------------------------------------------
   na::NAMapper mapper(
       arch, na::Configuration(
@@ -862,7 +882,7 @@ TEST(NAMapper, QAOA16Narrow) {
       ]
   }
   )");
-  std::stringstream  gridSS;
+  std::stringstream gridSS;
   gridSS << "x,y\n";
   // entangling zone (4 x 36 = 144 sites)
   for (std::size_t y = 0; y <= 36; y += 12) {
@@ -982,8 +1002,244 @@ ry(0.3223291) q;
 cp(pi) q[9],q[11];
 ry(-2.2154814) q;
 ry(2.2154814) q;)";
-  const auto&       circ = qc::QuantumComputation::fromQASM(qasm);
-  const auto&       arch = na::Architecture(archIS, gridSS);
+  const auto& circ = qc::QuantumComputation::fromQASM(qasm);
+  const auto& arch = na::Architecture(archIS, gridSS);
+  // ---------------------------------------------------------------------
+  na::NAMapper mapper(
+      arch, na::Configuration(
+                3, 2, na::NAMappingMethod::MaximizeParallelismHeuristic));
+  mapper.map(circ);
+  std::ignore = mapper.getStats();
+  EXPECT_TRUE(na::validateAODConstraints(mapper.getResult()));
+}
+
+TEST(NAMapper, QAOA16NarrowEntangling) {
+  std::istringstream archIS(R"({
+      "name": "Nature",
+      "initialZones": [
+          "storage"
+      ],
+      "zones": [
+          {
+              "name": "entangling",
+              "xmin": -300,
+              "xmax": 656,
+              "ymin": -10,
+              "ymax": 46,
+              "fidelity": 0.9959
+          },
+          {
+              "name": "storage",
+              "xmin": -300,
+              "xmax": 656,
+              "ymin": 47,
+              "ymax": 421,
+              "fidelity": 1
+          },
+          {
+              "name": "readout",
+              "xmin": -300,
+              "xmax": 656,
+              "ymin": 422,
+              "ymax": 456,
+              "fidelity": 0.99
+          }
+      ],
+      "operations": [
+          {
+              "name": "rz",
+              "type": "local",
+              "zones": [
+                  "entangling",
+                  "storage",
+                  "readout"
+              ],
+              "time": 0.5,
+              "fidelity": 0.999
+          },
+          {
+              "name": "ry",
+              "type": "global",
+              "zones": [
+                  "entangling",
+                  "storage",
+                  "readout"
+              ],
+              "time": 0.5,
+              "fidelity": 0.999
+          },
+          {
+              "name": "cz",
+              "type": "global",
+              "zones": [
+                  "entangling"
+              ],
+              "time": 0.2,
+              "fidelity": 0.9959
+          },
+          {
+              "name": "measure",
+              "type": "global",
+              "zones": [
+                  "readout"
+              ],
+              "time": 0.2,
+              "fidelity": 0.95
+          }
+      ],
+      "decoherence": {
+          "t1": 100000000,
+          "t2": 1500000
+      },
+      "interactionRadius": 2,
+      "noInteractionRadius": 5,
+      "minAtomDistance": 1,
+      "shuttling": [
+          {
+              "rows": 5,
+              "columns": 5,
+              "xmin": -2.5,
+              "xmax": 2.5,
+              "ymin": -2.5,
+              "ymax": 2.5,
+              "move": {
+                  "speed": 0.55,
+                  "fidelity": 1
+              },
+              "load": {
+                  "time": 20,
+                  "fidelity": 1
+              },
+              "store": {
+                  "time": 20,
+                  "fidelity": 1
+              }
+          }
+      ]
+  }
+  )");
+  std::stringstream gridSS;
+  gridSS << "x,y\n";
+  // entangling zone (4 x 36 = 144 sites)
+  for (std::size_t y = 0; y <= 36; y += 12) {
+    for (std::size_t x = 3; x <= 53; x += 10) {
+      gridSS << x << "," << y << "\n";
+    }
+  }
+  // storage zone (72 x 12 = 864 sites)
+  for (std::size_t y = 56; y <= 411; y += 5) {
+    for (std::size_t x = 0; x <= 55; x += 5) {
+      gridSS << x << "," << y << "\n";
+    }
+  }
+  // readout zone (4 x 12 = 48 sites)
+  for (std::size_t y = 431; y <= 446; y += 5) {
+    for (std::size_t x = 0; x <= 55; x += 5) {
+      gridSS << x << "," << y << "\n";
+    }
+  }
+  // total: 1056 sites
+  // For the test, we removed all rz gates because the mapping task remains the
+  // same
+  const std::string qasm = R"(OPENQASM 2.0;
+include "qelib1.inc";
+qreg q[16];
+ry(-pi/4) q;
+ry(pi/4) q;
+cp(pi) q[0],q[2];
+cp(pi) q[1],q[7];
+cp(pi) q[8],q[3];
+cp(pi) q[12],q[6];
+ry(-pi/2) q;
+ry(pi/2) q;
+cp(pi) q[0],q[2];
+cp(pi) q[1],q[7];
+cp(pi) q[8],q[3];
+cp(pi) q[12],q[6];
+ry(-pi/2) q;
+ry(pi/2) q;
+cp(pi) q[0],q[4];
+cp(pi) q[8],q[9];
+cp(pi) q[1],q[10];
+cp(pi) q[13],q[6];
+cp(pi) q[2],q[14];
+cp(pi) q[3],q[15];
+ry(-pi/2) q;
+ry(pi/2) q;
+cp(pi) q[0],q[4];
+cp(pi) q[8],q[9];
+cp(pi) q[1],q[10];
+cp(pi) q[13],q[6];
+cp(pi) q[2],q[14];
+cp(pi) q[3],q[15];
+ry(-pi/4) q;
+ry(pi/4) q;
+cp(pi) q[4],q[5];
+cp(pi) q[12],q[13];
+cp(pi) q[0],q[2];
+cp(pi) q[14],q[7];
+cp(pi) q[10],q[15];
+cp(pi) q[8],q[3];
+ry(-pi/2) q;
+ry(pi/2) q;
+cp(pi) q[4],q[5];
+cp(pi) q[12],q[13];
+cp(pi) q[0],q[2];
+cp(pi) q[14],q[7];
+cp(pi) q[10],q[15];
+cp(pi) q[8],q[3];
+ry(-pi/4) q;
+ry(pi/4) q;
+cp(pi) q[11],q[5];
+cp(pi) q[12],q[6];
+cp(pi) q[13],q[6];
+cp(pi) q[0],q[4];
+cp(pi) q[2],q[14];
+cp(pi) q[1],q[7];
+cp(pi) q[1],q[10];
+cp(pi) q[3],q[15];
+ry(-pi/2) q;
+ry(pi/2) q;
+cp(pi) q[11],q[5];
+cp(pi) q[12],q[6];
+cp(pi) q[13],q[6];
+cp(pi) q[0],q[4];
+cp(pi) q[2],q[14];
+cp(pi) q[1],q[7];
+cp(pi) q[1],q[10];
+cp(pi) q[3],q[15];
+ry(-pi/4) q;
+ry(pi/4) q;
+cp(pi) q[9],q[11];
+cp(pi) q[12],q[13];
+cp(pi) q[4],q[5];
+cp(pi) q[14],q[7];
+cp(pi) q[10],q[15];
+ry(-0.64469806) q;
+ry(0.64469806) q;
+cp(pi) q[9],q[11];
+cp(pi) q[12],q[13];
+cp(pi) q[4],q[5];
+cp(pi) q[14],q[7];
+cp(pi) q[10],q[15];
+ry(-2.2154814) q;
+ry(2.2154814) q;
+cp(pi) q[11],q[5];
+cp(pi) q[8],q[9];
+ry(-0.3223291) q;
+ry(0.3223291) q;
+cp(pi) q[11],q[5];
+cp(pi) q[8],q[9];
+ry(-pi/4) q;
+ry(pi/4) q;
+cp(pi) q[9],q[11];
+ry(-0.3223291) q;
+ry(0.3223291) q;
+cp(pi) q[9],q[11];
+ry(-2.2154814) q;
+ry(2.2154814) q;)";
+  const auto& circ = qc::QuantumComputation::fromQASM(qasm);
+  const auto& arch = na::Architecture(archIS, gridSS);
   // ---------------------------------------------------------------------
   na::NAMapper mapper(
       arch, na::Configuration(
