@@ -20,12 +20,12 @@ from mqt.qmap import (
     LookaheadHeuristic,
     Method,
     SwapReduction,
-    compile,
+    compile,  # noqa: A004
 )
 from mqt.qmap.visualization import SearchVisualizer
 
 
-@pytest.fixture()
+@pytest.fixture
 def example_circuit() -> QuantumCircuit:
     """Return a simple example circuit."""
     qc = QuantumCircuit(3)
